# jws/src/main/java/fr/epita/assistants/jws/presentation/rest/response/GameListResponse.java
package fr.epita.assistants.jws.presentation.rest.response;

import fr.epita.assistants.jws.utils.GameState;

public class GameListResponse {
    public Long id;
    public int players;
    public String state;

}